Replace int iterators with size_t for shift_rows

+++ src/kernel/tty.c
@@ -73,13 +73,13 @@ void term_put_at(size_t y, size_t x, uint8_t col, char c) {
 /* shift_rows: scrolls the terminal n rows */
 void shift_rows(int n) {
     /* Shift n rows */
-    for (int y = 0; y < VGA_HEIGHT - n; y++)
-        for (int x = 0; x < VGA_WIDTH; x++)
+    for (size_t y = 0; y < VGA_HEIGHT - n; y++)
+        for (size_t x = 0; x < VGA_WIDTH; x++)
             term_buf[y * VGA_WIDTH + x] = term_buf[(y + n) * VGA_WIDTH + x];
 
     /* Clear last n rows with the current color (only the background matters) */
-    for (int y = VGA_HEIGHT - n; y < VGA_HEIGHT; y++)
-        for (int x = 0; x < VGA_WIDTH; x++)
+    for (size_t y = VGA_HEIGHT - n; y < VGA_HEIGHT; y++)
+        for (size_t x = 0; x < VGA_WIDTH; x++)
             term_buf[y * VGA_WIDTH + x] = vga_entry(' ', term_col);
 }
